fix warning unused import on linux
